+++ src/main/java/chatbot/Peridot.java
@@ -1,3 +1,6 @@
+package chatbot;
+
+import storage.Storage;
 import tasks.TaskList;
 
 public class Peridot {
