+++ src/main/java/storage/Storage.java
@@ -1,16 +1,18 @@
-import tasks.Deadline;
-import tasks.Event;
-import tasks.Task;
-import tasks.Todo;
+package storage;
 
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileWriter;
 import java.io.IOException;
+import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Scanner;
-import java.time.LocalDate;
+
+import tasks.Deadline;
+import tasks.Event;
+import tasks.Task;
+import tasks.Todo;
 
 public class Storage {
 
@@ -36,27 +38,30 @@ public class Storage {
                 char taskType = line.charAt(0);
                 boolean done = false;
                 switch (taskType) {
-                    case 'T':
-                        done = line.charAt(1) == '1';
-                        loadedData.add(new Todo(line.substring(2), done));
-                        break;
-                    case 'D':
-                        done = line.charAt(1) == '1';
-                        String deadline = line.substring(line.indexOf('%') + 1);
-                        loadedData.add(new Deadline(line.substring(2, line.indexOf('%')),
-                                LocalDate.parse(deadline), done));
-                        break;
-                    case 'E':
-                        done = line.charAt(1) == '1';
-                        String start = line.substring(line.indexOf('%') + 1, line.indexOf('|'));
-                        String end = line.substring(line.indexOf('|') + 1);
-                        loadedData.add(new Event(line.substring(2, line.indexOf('%')),
-                                LocalDate.parse(start), LocalDate.parse(end), done));
+                case 'T':
+                    done = line.charAt(1) == '1';
+                    loadedData.add(new Todo(line.substring(2), done));
+                    break;
+                case 'D':
+                    done = line.charAt(1) == '1';
+                    String deadline = line.substring(line.indexOf('%') + 1);
+                    loadedData.add(new Deadline(line.substring(2, line.indexOf('%')),
+                            LocalDate.parse(deadline), done));
+                    break;
+                case 'E':
+                    done = line.charAt(1) == '1';
+                    String start = line.substring(line.indexOf('%') + 1, line.indexOf('|'));
+                    String end = line.substring(line.indexOf('|') + 1);
+                    loadedData.add(new Event(line.substring(2, line.indexOf('%')),
+                            LocalDate.parse(start), LocalDate.parse(end), done));
+                    break;
+                default:
+                    break;
                 }
             }
             return loadedData;
         } catch (FileNotFoundException e) {
-            System.out.println("Scanner failed in load() method of Storage-type object");
+            System.out.println("Scanner failed in load() method of storage.Storage-type object");
             return null;
         }
     }
