+++ src/test/java/tasks/TodoTest.java
@@ -1,8 +1,8 @@
 package tasks;
 
-import org.junit.jupiter.api.Test;
-
 import static org.junit.jupiter.api.Assertions.assertEquals;
+
+import org.junit.jupiter.api.Test;
 
 public class TodoTest {
     @Test
